requirements.txt
flake8
black
django
django-allauth
django-debug-toolbar
django-crispy-forms
django-filter
crispy-bootstrap5
Pillow